# docs/openapi/openapi.json
{"openapi": "3.0.2", "info": {"title": "interLink sidecar", "description": "openapi spec for interLink apis <-> provider sidecar communication", "version": "v0.0.0"}, "paths": {"/create": {"post": {"summary": "Create Pod", "operationId": "create_pod_create_post", "requestBody": {"content": {"application/json": {"schema": {"title": "Pods", "type": "array", "items": {"$ref": "#/components/schemas/Pod"}}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/delete": {"post": {"summary": "Delete Pod", "operationId": "delete_pod_delete_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PodRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/status": {"get": {"summary": "Status Pod", "operationId": "status_pod_status_get", "requestBody": {"content": {"application/json": {"schema": {"title": "Pods", "type": "array", "items": {"$ref": "#/components/schemas/PodRequest"}}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/getLogs": {"post": {"summary": "Get Logs", "operationId": "get_logs_getLogs_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LogRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ConfigMap": {"title": "ConfigMap", "required": ["metadata", "data"], "type": "object", "properties": {"metadata": {"$ref": "#/components/schemas/Metadata"}, "data": {"title": "Data", "type": "object"}}}, "ConfigMapSource": {"title": "ConfigMapSource", "required": ["configMapName", "items"], "type": "object", "properties": {"configMapName": {"title": "Configmapname", "type": "string"}, "items": {"title": "Items", "type": "array", "items": {"type": "object"}}}}, "Container": {"title": "Container", "required": ["name", "image", "command", "args", "resources"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "image": {"title": "Image", "type": "string"}, "tag": {"title": "Tag", "type": "string", "default": "latest"}, "command": {"title": "Command", "type": "array", "items": {"type": "string"}}, "args": {"title": "Args", "type": "array", "items": {"type": "string"}}, "resources": {"title": "Resources", "type": "object"}, "volumeMounts": {"title": "Volumemounts", "type": "array", "items": {"$ref": "#/components/schemas/VolumeMount"}}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "LogOpts": {"title": "LogOpts", "required": ["Tail", "Timestamps", "Previous", "SinceSeconds", "SinceTime"], "type": "object", "properties": {"Tail": {"title": "Tail", "type": "integer"}, "LimitBytes": {"title": "Limitbytes", "type": "integer"}, "Timestamps": {"title": "Timestamps", "type": "boolean"}, "Previous": {"title": "Previous", "type": "boolean"}, "SinceSeconds": {"title": "Sinceseconds", "type": "integer"}, "SinceTime": {"title": "Sincetime", "type": "string", "format": "date-time"}}}, "LogRequest": {"title": "LogRequest", "required": ["Namespace", "PodUID", "PodName", "ContainerName", "Opts"], "type": "object", "properties": {"Namespace": {"title": "Namespace", "type": "string"}, "PodUID": {"title": "Poduid", "type": "string"}, "PodName": {"title": "Podname", "type": "string"}, "ContainerName": {"title": "Containername", "type": "string"}, "Opts": {"$ref": "#/components/schemas/LogOpts"}}}, "Metadata": {"title": "Metadata", "required": ["name", "namespace", "uid", "annotations"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "namespace": {"title": "Namespace", "type": "string"}, "uid": {"title": "Uid", "type": "string"}, "annotations": {"title": "Annotations", "type": "object"}}}, "Pod": {"title": "Pod", "required": ["pod", "container"], "type": "object", "properties": {"pod": {"$ref": "#/components/schemas/PodRequest"}, "container": {"title": "Container", "type": "array", "items": {"$ref": "#/components/schemas/Volume"}}}}, "PodRequest": {"title": "PodRequest", "required": ["metadata", "spec"], "type": "object", "properties": {"metadata": {"$ref": "#/components/schemas/Metadata"}, "spec": {"$ref": "#/components/schemas/PodSpec"}}}, "PodSpec": {"title": "PodSpec", "required": ["containers"], "type": "object", "properties": {"containers": {"title": "Containers", "type": "array", "items": {"$ref": "#/components/schemas/Container"}}, "initContainers": {"title": "Initcontainers", "type": "array", "items": {"$ref": "#/components/schemas/Container"}}, "volumes": {"title": "Volumes", "type": "array", "items": {"$ref": "#/components/schemas/PodVolume"}}}}, "PodVolume": {"title": "PodVolume", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "volumeSource": {"$ref": "#/components/schemas/VolumeSource"}}}, "Secret": {"title": "Secret", "required": ["metadata", "data"], "type": "object", "properties": {"metadata": {"$ref": "#/components/schemas/Metadata"}, "data": {"title": "Data", "type": "object"}}}, "SecretSource": {"title": "SecretSource", "required": ["secretName", "items"], "type": "object", "properties": {"secretName": {"title": "Secretname", "type": "string"}, "items": {"title": "Items", "type": "array", "items": {"type": "object"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}, "Volume": {"title": "Volume", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "configMaps": {"title": "Configmaps", "type": "array", "items": {"$ref": "#/components/schemas/ConfigMap"}}, "secrets": {"title": "Secrets", "type": "array", "items": {"$ref": "#/components/schemas/Secret"}}, "emptyDirs": {"title": "Emptydirs", "type": "array", "items": {"type": "string"}}}}, "VolumeMount": {"title": "VolumeMount", "required": ["name", "mountPath"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "mountPath": {"title": "Mountpath", "type": "string"}, "subPath": {"title": "Subpath", "type": "string"}}}, "VolumeSource": {"title": "VolumeSource", "type": "object", "properties": {"emptyDir": {"title": "Emptydir", "type": "object"}, "secret": {"$ref": "#/components/schemas/SecretSource"}, "configMap": {"$ref": "#/components/schemas/ConfigMapSource"}}}}}}